Added Online Doc Link to README
Linked to the Github Pages documentation directly from the Readme, with context that the Readme documentation is an abridged version of what can be found there.

Update Version File

Added the increment to the version file to reflect the change to the documentation that is added to the master branch.