Fix cyclic plotting issue
The figure canvas instance linking events to actions according to the documentation is a very tenuous class instance without being explicitly stored in memory. It is very quickly garbage collected. This appears to be the reason that the cyclic plotting commands after the on_press. Explicitly storing the connection id created with mpl_connect appears to have circumvented this.