Run checks in single bash directive within azure-pipelines.yml file
Note that all commands are contained in a single bash directive. This
will produce CommandNotFoundError if the conda installation is not contained.

See:
        https://docs.microsoft.com/en-us/azure/devops/pipelines/languages/anaconda?view=azure-devops&tabs=ubuntu-16-04

Furthermore, "source $HOME/miniconda/etc/profile.d/conda.sh" is required
to allow for Azure to find "conda activate" commands. See below for
error messages:

Fixing CommandNotFoundError

CommandNotFoundError: Your shell has not been properly configured to use
'conda activate'.
If your shell is Bash or a Bourne variant, enable conda for the current
user with

    $ echo ". /usr/share/miniconda/etc/profile.d/conda.sh" >> ~/.bashrc

or, for all users, enable conda with

    $ sudo ln -s /usr/share/miniconda/etc/profile.d/conda.sh
/etc/profile.d/conda.sh

The options above will permanently enable the 'conda' command, but they
do NOT
put conda's base (root) environment on PATH.  To do so, run

    $ conda activate

in your terminal, or to put the base environment on PATH permanently,
run

    $ echo "conda activate" >> ~/.bashrc

Previous to conda 4.4, the recommended way to activate conda was to
modify PATH in
your ~/.bashrc file.  You should manually remove the line that looks
like

    export PATH="/usr/share/miniconda/bin:$PATH"