Adding Github ISSUE/PR templates
These files allow for templating of issues and pull requests on Github
for easier contributions.

The layout is akin to:
	https://github.com/tallamjr/dotfiles/issues/new/choose
where one chooses an issue option, BUG, FEATURE_REQUEST, or just make a
general issue.

In addition, a PULL_REQUEST_TEMPLATE has been created to allow for
consistent documentation of what a PR achieves.

Fixes #126