Adding utils to PYTHONPATH via setup.py file
As mentioned in PR#165, files such as plasticc_utils.py should be
accessible from time of install of snmachine. The snippet added for
PR165 was incompatible for snclassifer.py and thus needed updating.

The changes in this commit add 'utils' as an included package of 'snmachine'
where one can import functions with the following:

>>> from utils import plasticc_utils
>>> plasticc_utils.plasticc_log_loss( ...

REF:

    https://github.com/LSSTDESC/snmachine/pull/165#discussion_r291628870

    modified:   setup.py
    modified:   snmachine/snclassifier.py
    modified:   snmachine/version.py
    new file:   utils/__init__.py