Change from method_directory -> analysis_directory
This is to be consistent with other variable names in the script.

[SQUASH]
Adding default option set to false for saving the subset list.

Not saving is preferred for quick analysis as it will load in a file already
on disk if it exists which would mean exploration analysis is hindered.