Add support for ELAsTiCC's training dataset
The code (in `_training_data.py`) implements the `ElasticcTrainingData`
class, modeled on the classes implemented in `sndata.py`. Unlike the
`PlasticcData` class (and the other extant similar classes), this new
class does not inherit from `EmptyDataset`; this might change in the
future.

The code is implemented with interactive usage in mind. Hence
`__init__.py` is responsible for constructing the effective module's
namespace, and the non-public implication for `_training_data.py` and
`_training_metadata.py'. (The latter stores the metadata for the
training data files themselves.)