Make intent for sncosmo compatibility explicit
While the Tables produced aren't yet fully compatible with sncosmo
(they're missing the zp and zpsys columns), this at least makes the
intent to do so explicit. The values chosen here appear in sncosmo's
documentation as valid aliases for the keys, and align with the values
used elsewhere in snmachine. In principle, the SNCOSMO_COLS dictionary
should be defined higher up in snmachine, such that it can be applied
package wide.